refactor(dashboard): make types.ts a pure barrel over the workstation shards
A refactor had restored the full 11k-line body of types.ts while keeping the
shard re-exports, declaring 1,018 type names twice; the local copies
silently shadowed the shards and had already diverged. The diverged
definitions (StatementImportCommitResult, AccountingReportingSummary — the
versions the app compiles against) are folded into workstation-4, the five
types that existed only in types.ts move to workstation-4, and types.ts is
reduced to re-exports only. Exported-name parity verified before/after
(1,048 names, identical sets); tsc, strict-null typecheck, and the vitest
suite pass.
